feat: add types to login page
