+++ src/pages/login/interface.ts
@@ -1,0 +1,4 @@
+export interface ILoginState {
+  email: string;
+  password: string;
+}
